Some code and layout refactor
