+++ app/src/main/res/layout/fragment_media_favourites.xml
@@ -6,15 +6,6 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent" >
 
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/rv_track"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_marginTop="@dimen/margin_base_16"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
     <ImageView
         android:id="@+id/img_placeholder"
         android:layout_width="wrap_content"
@@ -22,7 +13,7 @@
         android:layout_marginTop="106dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/rv_track"
+        app:layout_constraintTop_toTopOf="parent"
         tools:src="@drawable/img_search_error" />
 
     <TextView
@@ -32,5 +23,15 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/img_placeholder"
         tools:text="Ваша медиатека пуста" />
+
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/rv_track"
+        android:layout_width="match_parent"
+        android:layout_height="0dp"
+        android:layout_marginTop="@dimen/margin_base_16"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/tv_placeholder"
+        app:layout_constraintBottom_toBottomOf="parent"/>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
