+++ app/src/main/res/layout/fragment_media_playlists.xml
@@ -39,6 +39,7 @@
         android:layout_marginHorizontal="16dp"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/tv_placeholder" />
+        app:layout_constraintTop_toBottomOf="@id/tv_placeholder"
+        app:layout_constraintBottom_toBottomOf="parent" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
